# backend/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Card": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"scryfallId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"manaCost": {
						"type": "string"
					},
					"oracleText": {
						"type": "string"
					},
					"power": {
						"type": "number",
						"format": "double"
					},
					"toughness": {
						"type": "number",
						"format": "double"
					},
					"set": {
						"type": "string"
					},
					"cmc": {
						"type": "number",
						"format": "double"
					},
					"colors": {
						"items": {
							"$ref": "#/components/schemas/Color"
						},
						"type": "array"
					},
					"type": {
						"type": "string"
					},
					"faces": {
						"items": {
							"$ref": "#/components/schemas/CardFace"
						},
						"type": "array"
					},
					"tokens": {
						"items": {
							"$ref": "#/components/schemas/Token"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"scryfallId",
					"name",
					"manaCost",
					"oracleText",
					"power",
					"toughness",
					"set",
					"cmc",
					"colors",
					"type",
					"faces",
					"tokens"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"loggedIn": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"staff": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"admin": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "motticon-backend",
		"version": "0.0.1",
		"description": "MottiCon backend",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
...
		"/card/name/{cardname}": {
			"get": {
				"operationId": "GetCardByName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Card"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cardname",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/cube/tournament/{tournamentId}": {
			"get": {
				"operationId": "GetCubesForTournament",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CubeDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/match/submitResult": {
			"post": {
				"operationId": "SubmitResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MatchDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"player2GamesWon": {
										"type": "number",
										"format": "double"
									},
									"player1GamesWon": {
										"type": "number",
										"format": "double"
									},
									"resultSubmittedBy": {
										"type": "number",
										"format": "double"
									},
									"roundId": {
										"type": "number",
										"format": "double"
									},
									"matchId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"player2GamesWon",
									"player1GamesWon",
									"resultSubmittedBy",
									"roundId",
									"matchId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/tournament/{tournamentId}/round/{roundId}/results": {
			"get": {
				"operationId": "GenerateCsvFromRound",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "roundId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament": {
			"get": {
				"operationId": "GetAllTournaments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TournamentDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": []
			}
		},
		"/tournament/future": {
			"get": {
				"operationId": "GetFutureTournaments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TournamentDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": []
			}
		},
		"/tournament/past": {
			"get": {
				"operationId": "GetPastTournaments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TournamentDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": []
			}
		},
		"/tournament/ongoing": {
			"get": {
				"operationId": "GetOngoingTournaments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TournamentDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": []
			}
		},
		"/tournament/{tournamentId}": {
			"get": {
				"operationId": "GetTournament",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/enrollment": {
			"get": {
				"operationId": "GetTournamentEnrollments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/drafts": {
			"get": {
				"operationId": "GetTournamentAndDrafts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/draft": {
			"get": {
				"operationId": "GetCurrentDraft",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DraftDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/round": {
			"get": {
				"operationId": "GetCurrentRound",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoundDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{_tournamentId}/round/{roundId}/match/{playerId}": {
			"get": {
				"operationId": "GetMatch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MatchDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "_tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "roundId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "playerId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/round/recent": {
			"get": {
				"operationId": "GetMostRecentRound",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoundDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/standings/{roundNumber}": {
			"get": {
				"operationId": "GetStandings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/StandingsRow"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "roundNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/score/{userId}": {
			"get": {
				"operationId": "GetPreviousScore",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayerTournamentScoreDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{id}/cubes": {
			"get": {
				"operationId": "GetCubesForTournament",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CubeDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/enroll/{userId}": {
			"post": {
				"operationId": "EnrollIntoTournament",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/cancel/{userId}": {
			"post": {
				"operationId": "CancelEnrollment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/drop/{userId}": {
			"post": {
				"operationId": "DropFromTournament",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/preferences": {
			"post": {
				"operationId": "SetCubePreferences",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/tournament/preferences/delete": {
			"put": {
				"operationId": "DeleteCubePreferences",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/tournament/{tournamentId}/preferences/{userId}": {
			"get": {
				"operationId": "GetPreferencesForUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Preference"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"loggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/start": {
			"put": {
				"operationId": "StartTournament",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"staff": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/end": {
			"put": {
				"operationId": "EndTournament",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"staff": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/draft/generate": {
			"post": {
				"operationId": "GenerateDrafts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"staff": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/draft/{draftId}/initiate": {
			"put": {
				"operationId": "InitiateDraft",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DraftDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"staff": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "draftId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/draft/{draftId}/start": {
			"put": {
				"operationId": "StartDraft",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DraftDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"staff": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "draftId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tournament/{tournamentId}/draft/{draftId}/end": {
			"put": {
				"operationId": "EndDraft",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"staff": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tournamentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "draftId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/tournament/create": {
			"post": {
				"operationId": "CreateTournament",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}